Extract & rename transform animation import
Same as export...
