+++ io_scene_nif/modules/animation/animation_import.py
@@ -37,35 +37,13 @@
 #
 # ***** END LICENSE BLOCK *****
 import bpy
-import mathutils
-
-from bisect import bisect_left
 
 from pyffi.formats.nif import NifFormat
 
-from io_scene_nif.modules import armature
+from io_scene_nif.modules.animation.transform_import import TransformAnimation
 from io_scene_nif.utility import nif_utils
 from io_scene_nif.utility.util_global import NifOp
 from io_scene_nif.utility.util_logging import NifLog
-
-
-# TODO [animation][util] interpolate() should perhaps be moved to utils?
-def interpolate(x_out, x_in, y_in):
-    """
-    sample (x_in I y_in) at x coordinates x_out
-    """
-    y_out = []
-    intervals = zip(x_in, x_in[1:], y_in, y_in[1:])
-    slopes = [(y2 - y1) / (x2 - x1) for x1, x2, y1, y2 in intervals]
-    # if we had just one input, slope will be 0 for constant extrapolation
-    if not slopes:
-        slopes = [0, ]
-    for x in x_out:
-        i = bisect_left(x_in, x) - 1
-        # clamp to valid range
-        i = max(min(i, len(slopes) - 1), 0)
-        y_out.append(y_in[i] + slopes[i] * (x - x_in[i]))
-    return y_out
 
 
 class Animation:
@@ -73,7 +51,7 @@ class Animation:
     def __init__(self):
         self.object_animation = ObjectAnimation(self)
         self.material_animation = MaterialAnimation(self)
-        self.armature_animation = ArmatureAnimation(self)
+        self.transform = TransformAnimation(self)
         # set a dummy here, update via set_frames_per_second
         self.fps = 30
 
@@ -389,181 +367,3 @@ class MaterialAnimation:
                             else:
                                 self.animationhelper.add_key(fcurves, key.time, (key.value,), interp)
 
-
-class ArmatureAnimation:
-
-    def __init__(self, parent):
-        self.animationhelper = parent
-
-    def import_kf_standalone(self, kf_root, b_armature_obj, bind_data):
-        """Import a kf animation. Needs a suitable armature in blender scene."""
-
-        NifLog.info("Importing KF tree")
-
-        # check that this is an Oblivion style kf file
-        if not isinstance(kf_root, NifFormat.NiControllerSequence):
-            raise nif_utils.NifError("non-Oblivion .kf import not supported")
-
-        # import text keys
-        self.animationhelper.import_text_keys(kf_root)
-
-        b_action = self.animationhelper.create_action(b_armature_obj, kf_root.name.decode())
-        # go over all controlled blocks (NiKeyframeController)
-        for controlledblock in kf_root.controlled_blocks:
-            # get bone name
-            # ZT2 - old way is not supported by pyffi's get_node_name()
-            n_name = controlledblock.target_name
-            # fallout (node_name) & Loki (StringPalette)
-            if not n_name:
-                n_name = controlledblock.get_node_name()
-            bone_name = armature.get_bone_name_for_blender(n_name)
-            if bone_name not in b_armature_obj.data.bones:
-                continue
-            b_bone = b_armature_obj.data.bones[bone_name]
-            # import bone priority
-            b_bone.niftools.bonepriority = controlledblock.priority
-            # import animation
-            if bone_name in bind_data:
-                niBone_bind_scale, niBone_bind_rot_inv, niBone_bind_trans = bind_data[bone_name]
-                # ZT2
-                kfc = controlledblock.controller
-                # fallout, Loki
-                if not kfc:
-                    kfc = controlledblock.interpolator
-                if kfc:
-                    self.import_keyframe_controller(kfc, b_armature_obj, bone_name, niBone_bind_scale,
-                                                    niBone_bind_rot_inv, niBone_bind_trans)
-        # fallout: set global extrapolation mode here (older versions have extrapolation per controller)
-        if kf_root.cycle_type:
-            extend = self.animationhelper.get_extend_from_cycle_type(kf_root.cycle_type)
-            self.animationhelper.set_extrapolation(extend, b_action.fcurves)
-
-    # TODO [animation] Is scale param required or can be removed, not used
-    def import_keyframe_controller(self, n_kfc, b_obj, bone_name=None, n_bone_bind_scale=None, n_bone_bind_rot_inv=None, n_bone_bind_trans=None):
-        b_action = b_obj.animation_data.action
-
-        if bone_name:
-            b_obj = b_obj.pose.bones[bone_name]
-
-        translations = []
-        scales = []
-        rotations = []
-        eulers = []
-        n_kfd = None
-
-        # B-spline curve import
-        if isinstance(n_kfc, NifFormat.NiBSplineInterpolator):
-            # used by WLP2 (tiger.kf), but only for non-LocRotScale data
-            # eg. bone stretching - see controlledblock.get_variable_1()
-            # do not support this for now, no good representation in Blender
-            if isinstance(n_kfc, NifFormat.NiBSplineCompFloatInterpolator):
-                # pyffi lacks support for this, but the following gets float keys
-                # keys = list(kfc._getCompKeys(kfc.offset, 1, kfc.bias, kfc.multiplier))
-                return
-            times = list(n_kfc.get_times())
-            # just do these temp steps to avoid generating empty fcurves down the line
-            trans_temp = [mathutils.Vector(tup) for tup in n_kfc.get_translations()]
-            if trans_temp:
-                translations = zip(times, trans_temp)
-            rot_temp = [mathutils.Quaternion(tup) for tup in n_kfc.get_rotations()]
-            if rot_temp:
-                rotations = zip(times, rot_temp)
-            scale_temp = list(n_kfc.get_scales())
-            if scale_temp:
-                scales = zip(times, scale_temp)
-            # Bsplines are Bezier curves
-            interp_rot = interp_loc = interp_scale = "BEZIER"
-        else:
-            # ZT2 & Fallout
-            n_kfd = n_kfc.data
-
-        # ZT2 - get extrapolation for every kfc
-        if isinstance(n_kfc, NifFormat.NiKeyframeController):
-            flags = n_kfc.flags
-        # fallout, Loki - we set extrapolation according to the root NiControllerSequence.cycle_type
-        else:
-            flags = None
-        if isinstance(n_kfd, NifFormat.NiKeyframeData):
-            interp_rot = self.animationhelper.get_b_interp_from_n_interp(n_kfd.rotation_type)
-            interp_loc = self.animationhelper.get_b_interp_from_n_interp(n_kfd.translations.interpolation)
-            interp_scale = self.animationhelper.get_b_interp_from_n_interp(n_kfd.scales.interpolation)
-            if n_kfd.rotation_type == 4:
-                b_obj.rotation_mode = "XYZ"
-                # uses xyz rotation
-                if n_kfd.xyz_rotations[0].keys:
-                    # euler keys need not be sampled at the same time in KFs
-                    # but we need complete key sets to do the space conversion
-                    # so perform linear interpolation to import all keys properly
-
-                    # get all the keys' times
-                    times_x = [key.time for key in n_kfd.xyz_rotations[0].keys]
-                    times_y = [key.time for key in n_kfd.xyz_rotations[1].keys]
-                    times_z = [key.time for key in n_kfd.xyz_rotations[2].keys]
-                    # the unique time stamps we have to sample all curves at
-                    times_all = sorted(set(times_x + times_y + times_z))
-                    # the actual resampling
-                    x_r = interpolate(times_all, times_x, [key.value for key in n_kfd.xyz_rotations[0].keys])
-                    y_r = interpolate(times_all, times_y, [key.value for key in n_kfd.xyz_rotations[1].keys])
-                    z_r = interpolate(times_all, times_z, [key.value for key in n_kfd.xyz_rotations[2].keys])
-                eulers = zip(times_all, zip(x_r, y_r, z_r))
-            else:
-                b_obj.rotation_mode = "QUATERNION"
-                rotations = [(key.time, key.value) for key in n_kfd.quaternion_keys]
-
-            if n_kfd.scales.keys:
-                scales = [(key.time, key.value) for key in n_kfd.scales.keys]
-
-            if n_kfd.translations.keys:
-                translations = [(key.time, key.value) for key in n_kfd.translations.keys]
-
-        if eulers:
-            NifLog.debug('Rotation keys...(euler)')
-            fcurves = self.animationhelper.create_fcurves(b_action, "rotation_euler", range(3), flags, bone_name)
-            for t, val in eulers:
-                key = mathutils.Euler(val)
-                if bone_name:
-                    key = armature.import_keymat(n_bone_bind_rot_inv, key.to_matrix().to_4x4()).to_euler()
-                self.animationhelper.add_key(fcurves, t, key, interp_rot)
-        elif rotations:
-            NifLog.debug('Rotation keys...(quaternions)')
-            fcurves = self.animationhelper.create_fcurves(b_action, "rotation_quaternion", range(4), flags, bone_name)
-            for t, val in rotations:
-                key = mathutils.Quaternion([val.w, val.x, val.y, val.z])
-                if bone_name:
-                    key = armature.import_keymat(n_bone_bind_rot_inv, key.to_matrix().to_4x4()).to_quaternion()
-                self.animationhelper.add_key(fcurves, t, key, interp_rot)
-        if translations:
-            NifLog.debug('Translation keys...')
-            fcurves = self.animationhelper.create_fcurves(b_action, "location", range(3), flags, bone_name)
-            for t, val in translations:
-                key = mathutils.Vector([val.x, val.y, val.z])
-                if bone_name:
-                    key = armature.import_keymat(n_bone_bind_rot_inv,
-                                                 mathutils.Matrix.Translation(key - n_bone_bind_trans)).to_translation()
-                self.animationhelper.add_key(fcurves, t, key, interp_loc)
-        if scales:
-            NifLog.debug('Scale keys...')
-            fcurves = self.animationhelper.create_fcurves(b_action, "scale", range(3), flags, bone_name)
-            for t, val in scales:
-                key = (val, val, val)
-                self.animationhelper.add_key(fcurves, t, key, interp_scale)
-
-    def import_object_animation(self, n_block, b_obj):
-        """Loads an animation attached to a nif block (non-skeletal). Becomes the object level animation of the blender object."""
-        NifLog.debug('Importing animation for object %s'.format(b_obj.name))
-
-        n_kfc = nif_utils.find_controller(n_block, NifFormat.NiKeyframeController)
-        if n_kfc:
-            self.animationhelper.create_action(b_obj, b_obj.name + "-Anim")
-            self.import_keyframe_controller(n_kfc, b_obj)
-
-    def import_bone_animation(self, n_block, b_armature_obj, bone_name):
-        """Loads an animation attached to a nif block (skeletal). Becomes the pose bone level animation of the blender object."""
-        NifLog.debug('Importing animation for bone %s'.format(bone_name))
-
-        n_kfc = nif_utils.find_controller(n_block, NifFormat.NiKeyframeController)
-        if n_kfc:
-            bone_bm = nif_utils.import_matrix(n_block)  # base pose
-            n_bone_bind_scale, n_bone_bind_rot, n_bone_bind_trans = nif_utils.decompose_srt(bone_bm)
-            niBone_bind_rot_inv = n_bone_bind_rot.to_4x4().inverted()
-            self.import_keyframe_controller(n_kfc, b_armature_obj, bone_name, n_bone_bind_scale, niBone_bind_rot_inv, n_bone_bind_trans)
